refactor(cli): move the inspect group into its own module
Cuts `gpio inspect` out of cli/main.py into
geoparquet_io/cli/commands/inspect.py, unchanged, following the pattern set
by the sort, publish and process groups.

`InspectDefaultGroup` and `_handle_meta_display` travel with the group: each
has exactly one user, `inspect`. `_handle_meta_display` sat between the
`extract` and `add` groups in main.py, so it is the one piece here that was
not already adjacent to its group.

`tests/test_inspect.py` resolved `_validate_parquet_input` through
`geoparquet_io.cli.main` and is repointed at the new module.

tests/data/cli_surface.json is byte-identical.
